Break up schema module & refactor methods
