+++ backend/server/graphql/__init__.py
@@ -1,0 +1,6 @@
+"""Module for GraphQL schema and related queries and types."""
+
+import graphene
+from .schema import Query
+
+schema = graphene.Schema(query=Query)
