+++ backend/server/graphql/types/__init__.py
@@ -1,0 +1,4 @@
+"""GraphQL type classes."""
+
+from .season import SeasonType, RoundType
+from .models import PredictionType
