+++ backend/server/types.py
@@ -1,9 +1,12 @@
 """Contains custom mypy TypedDicts that are used across the application."""
 
-from typing import Union
+from typing import Union, List
 from datetime import datetime
+
 from mypy_extensions import TypedDict
 from pandas import Timestamp
+from django.db.models import QuerySet
+
 
 FixtureData = TypedDict(
     "FixtureData",
@@ -63,3 +66,24 @@ CleanPredictionData = TypedDict(
 )
 
 MlModel = TypedDict("MlModel", {"name": str, "filepath": str})
+
+ModelMetric = TypedDict(
+    "ModelMetric",
+    {
+        "ml_model__name": str,
+        "cumulative_correct_count": int,
+        "cumulative_accuracy": float,
+        "cumulative_mean_absolute_error": float,
+        "cumulative_margin_difference": int,
+        "cumulative_bits": float,
+    },
+)
+
+RoundPrediction = TypedDict(
+    "RoundPrediction",
+    {
+        "match__round_number": int,
+        "model_metrics": List[ModelMetric],
+        "matches": QuerySet,
+    },
+)
